requirements.txt
pandas
numpy
matplotlib
scikit-learn
python-dotenv
